[Refactor]: 로또 번호 검증 로직을 클래스로 분리 및 WinningNumber와 Lotto 클래스에 상속
