+++ src/WinningNumbers.js
@@ -1,30 +1,11 @@
-import { ERROR, LOTTO } from './constant/index.js';
+import LottoNumbers from './LottoNumbers.js';
 
-class WinningNumbers {
+class WinningNumbers extends LottoNumbers {
   #winningNumbers;
-  #bonusNumber;
 
   constructor(winningNumbers) {
-    WinningNumbers.#validate(winningNumbers);
+    super(winningNumbers);
     this.#winningNumbers = winningNumbers;
-  }
-
-  static #validate(numbers) {
-    if (numbers.length !== LOTTO.NUMBERS.LENGTH) {
-      throw new Error(ERROR.WINNING_NUMBERS.LENGTH);
-    }
-
-    const isInvalidRange = numbers.some(
-      number => number < LOTTO.NUMBERS.MIN || number > LOTTO.NUMBERS.MAX
-    );
-    if (isInvalidRange) {
-      throw new Error(ERROR.WINNING_NUMBERS.RANGE);
-    }
-
-    const uniqueNumbers = new Set(numbers);
-    if (uniqueNumbers.size !== 6) {
-      throw new Error(ERROR.WINNING_NUMBERS.UNIQE);
-    }
   }
 }
