+++ src/constant/error.js
@@ -15,10 +15,4 @@ const LOTTO_NUMBERS = Object.freeze({
   UNIQE: '[ERROR] 로또 번호는 중복이 없어야 합니다.',
 });
 
-const WINNING_NUMBERS = Object.freeze({
-  LENGTH: '[ERROR] 당첨 번호는 6개여야 합니다.',
-  RANGE: '[ERROR] 당첨 번호는 1에서 45 사이의 숫자여야 합니다.',
-  UNIQE: '[ERROR] 당첨 번호는 중복이 없어야 합니다.',
-});
-
-export { BONUS_NUMBER, BUYING_PRICE, LOTTO_NUMBERS, WINNING_NUMBERS };
+export { BONUS_NUMBER, BUYING_PRICE, LOTTO_NUMBERS };
