+++ src/constant/index.js
@@ -1,14 +1,9 @@
-import {
-  BONUS_NUMBER,
-  BUYING_PRICE,
-  LOTTO_NUMBERS,
-  WINNING_NUMBERS,
-} from './error.js';
+import { BONUS_NUMBER, BUYING_PRICE, LOTTO_NUMBERS } from './error.js';
 import { NUMBERS, PRICE } from './lotto.js';
 import INPUT from './input.js';
 import RANDOM from './random.js';
 
-const ERROR = { BONUS_NUMBER, BUYING_PRICE, LOTTO_NUMBERS, WINNING_NUMBERS };
+const ERROR = { BONUS_NUMBER, BUYING_PRICE, LOTTO_NUMBERS };
 const LOTTO = { NUMBERS, PRICE };
 
 export { ERROR, INPUT, LOTTO, RANDOM };
